+++ src/components/CurriculumComponents/UnitModal/UnitModal.tsx
@@ -16,6 +16,7 @@ import { notUndefined, Unit, YearData, Lesson } from "@/utils/curriculum/types";
 import useAnalytics from "@/context/Analytics/useAnalytics";
 import { ComponentTypeValueType } from "@/browser-lib/avo/Avo";
 import { getTitleFromSlug } from "@/fixtures/shared/helper";
+import { getIsUnitDescriptionEnabled } from "@/utils/curriculum/features";
 
 type UnitModalProps = {
   unitData: Unit | null;
@@ -74,7 +75,7 @@ const UnitModal: FC<UnitModalProps> = ({
   const yearTitle = unitData
     ? getYearGroupTitle(
         yearData,
-        unitData.actions?.programme_field_overrides?.Year ?? unitData.year,
+        unitData.actions?.programme_field_overrides?.year_slug ?? unitData.year,
       )
     : "";
 
@@ -100,6 +101,8 @@ const UnitModal: FC<UnitModalProps> = ({
       });
     }
   };
+
+  const isUnitDescriptionEnabled = getIsUnitDescriptionEnabled(unitData);
 
   return (
     <>
@@ -164,10 +167,7 @@ const UnitModal: FC<UnitModalProps> = ({
                     handleUnitOverviewExploredAnalytics
                   }
                   threads={unitData.threads}
-                  isUnitDescriptionEnabled={
-                    unitData.parent_programme_features?.unit_description ===
-                    true
-                  }
+                  isUnitDescriptionEnabled={isUnitDescriptionEnabled}
                   whyThisWhyNow={unitData.why_this_why_now}
                   description={unitData.description}
                   lessons={unitData.lessons}
@@ -246,8 +246,7 @@ const UnitModal: FC<UnitModalProps> = ({
                                 description: optionalUnit.description,
                                 whyThisWhyNow: optionalUnit.why_this_why_now,
                                 isUnitDescriptionEnabled:
-                                  unitData.parent_programme_features
-                                    ?.unit_description === true,
+                                  isUnitDescriptionEnabled,
                                 handleUnitOverviewExploredAnalytics:
                                   handleUnitOverviewExploredAnalytics,
                               });