+++ src/components/TeacherComponents/ListItemCard/ListItemCard.tsx
@@ -12,7 +12,7 @@ export type ListItemCardProps = {
   isHovered: boolean;
   children: React.ReactNode;
   background: OakColorName;
-  expired: boolean | null;
+  disabled: boolean | null;
   index: number;
   containerProps: {
     onClick: MouseEventHandler<HTMLDivElement>;
@@ -26,7 +26,13 @@ export type ListItemCardProps = {
  */
 
 const ListItemCard: FC<ListItemCardProps> = (props) => {
-  const { children, isHovered, containerProps, background, expired } = props;
+  const {
+    children,
+    isHovered,
+    containerProps,
+    background,
+    disabled: expired,
+  } = props;
 
   const applyHoverStyles = isHovered && !expired;