+++ src/node-lib/curriculum-api-2023/queries/teacherPreviewLessonListing/teacherPreviewLessonListing.query.ts
@@ -1,77 +1,56 @@
-import {
-  ProgrammeFields,
-  StaticLesson,
-} from "@oaknational/oak-curriculum-schema";
+import { ProgrammeFields } from "@oaknational/oak-curriculum-schema";
+
+import { BetaLessonListSchema } from "./teacherPreviewLessonListing.schema";
 
 import {
   lessonListingPageDataSchema,
   LessonListingPageData,
   partialSyntheticUnitvariantLessonsArraySchema,
   partialSyntheticUnitvariantLessonsSchema,
-} from "./lessonListing.schema";
-
+} from "@/node-lib/curriculum-api-2023/queries/lessonListing/lessonListing.schema";
 import { Sdk } from "@/node-lib/curriculum-api-2023/sdk";
 import OakError from "@/errors/OakError";
-import {
-  LessonListSchema,
-  Actions,
-  LessonListItem,
-} from "@/node-lib/curriculum-api-2023/shared.schema";
-import { LessonListingQuery } from "@/node-lib/curriculum-api-2023/generated/sdk";
+import { Actions } from "@/node-lib/curriculum-api-2023/shared.schema";
+import { TeacherPreviewLessonListingQuery } from "@/node-lib/curriculum-api-2023/generated/sdk";
 import { applyGenericOverridesAndExceptions } from "@/node-lib/curriculum-api-2023/helpers/overridesAndExceptions";
 import { getCorrectYear } from "@/node-lib/curriculum-api-2023/helpers/getCorrectYear";
 import { getIntersection } from "@/utils/getIntersection";
 import keysToCamelCase from "@/utils/snakeCaseConverter";
 
 export const getTransformedLessons = (
-  lessons: LessonListingQuery["lessons"],
-): LessonListSchema => {
-  return lessons[0]?.static_lesson_list
-    ?.sort((a: StaticLesson, b: StaticLesson) => a.order - b.order)
-    .map((staticLesson: StaticLesson) => {
-      const publishedLesson = lessons.find(
-        (lesson) => lesson.lesson_slug === staticLesson.slug,
-      );
+  lessons: TeacherPreviewLessonListingQuery["lessons"],
+): BetaLessonListSchema => {
+  return lessons
+    .map((l) => {
+      const lesson = partialSyntheticUnitvariantLessonsSchema.parse(l);
+      const hasCopyrightMaterial =
+        l.lesson_data.copyright_content?.find(
+          (c: { copyright_info: string }) =>
+            c.copyright_info === "This lesson contains copyright material.",
+        ) !== undefined;
 
-      if (publishedLesson) {
-        const lesson =
-          partialSyntheticUnitvariantLessonsSchema.parse(publishedLesson);
-        const hasCopyrightMaterial =
-          publishedLesson.lesson_data.copyright_content?.find(
-            (c: { copyright_info: string }) =>
-              c.copyright_info === "This lesson contains copyright material.",
-          ) !== undefined;
-
-        const transformedLesson = {
-          lessonSlug: lesson.lesson_slug,
-          lessonTitle: lesson.lesson_data.title,
-          description:
-            lesson.lesson_data.description ||
-            lesson.lesson_data.pupil_lesson_outcome,
-          pupilLessonOutcome: lesson.lesson_data.pupil_lesson_outcome,
-          expired: Boolean(lesson.lesson_data.deprecated_fields?.expired),
-          quizCount:
-            (lesson.lesson_data.quiz_id_starter ? 1 : 0) +
-            (lesson.lesson_data.quiz_id_exit ? 1 : 0),
-          videoCount: lesson.lesson_data.video_id ? 1 : 0,
-          presentationCount: lesson.lesson_data.asset_id_slidedeck ? 1 : 0,
-          worksheetCount: lesson.lesson_data.asset_id_worksheet ? 1 : 0,
-          hasCopyrightMaterial,
-          orderInUnit: lesson.order_in_unit,
-          lessonCohort: lesson.lesson_data._cohort,
-          actions: (keysToCamelCase(lesson.actions) || null) as Actions,
-          isUnpublished: false,
-        };
-        return transformedLesson;
-      } else {
-        return {
-          lessonSlug: staticLesson.slug,
-          lessonTitle: staticLesson.title,
-          orderInUnit: staticLesson.order,
-          isUnpublished: true,
-        };
-      }
-    });
+      const transformedLesson = {
+        lessonSlug: lesson.lesson_slug,
+        lessonTitle: lesson.lesson_data.title,
+        description:
+          lesson.lesson_data.description ||
+          lesson.lesson_data.pupil_lesson_outcome,
+        pupilLessonOutcome: lesson.lesson_data.pupil_lesson_outcome,
+        expired: Boolean(lesson.lesson_data.deprecated_fields?.expired),
+        quizCount:
+          (lesson.lesson_data.quiz_id_starter ? 1 : 0) +
+          (lesson.lesson_data.quiz_id_exit ? 1 : 0),
+        videoCount: lesson.lesson_data.video_id ? 1 : 0,
+        presentationCount: lesson.lesson_data.asset_id_slidedeck ? 1 : 0,
+        worksheetCount: lesson.lesson_data.asset_id_worksheet ? 1 : 0,
+        hasCopyrightMaterial,
+        orderInUnit: lesson.order_in_unit,
+        lessonCohort: lesson.lesson_data._cohort,
+        actions: (keysToCamelCase(lesson.actions) || null) as Actions,
+      };
+      return transformedLesson;
+    })
+    .sort((a, b) => a.orderInUnit - b.orderInUnit);
 };
 
 type PackagedUnitData = {
@@ -85,7 +64,7 @@ type PackagedUnitData = {
 
 export const getPackagedUnit = (
   packagedUnitData: PackagedUnitData,
-  unitLessons: LessonListSchema,
+  unitLessons: BetaLessonListSchema,
 ): LessonListingPageData => {
   const {
     programmeFields,
@@ -101,11 +80,9 @@ export const getPackagedUnit = (
     programmeFields,
   });
 
-  const combinedActions = getIntersection<LessonListItem["actions"]>(
-    unitLessons
-      .filter((lesson) => !lesson.isUnpublished)
-      .map((lesson) => lesson.actions),
-  ) as Actions;
+  const combinedActions = getIntersection<
+    BetaLessonListSchema[number]["actions"]
+  >(unitLessons.map((lesson) => lesson.actions)) as Actions;
 
   return {
     programmeSlug,
@@ -122,7 +99,10 @@ export const getPackagedUnit = (
     examBoardTitle: modifiedProgrammeFields.examboard,
     yearSlug: modifiedProgrammeFields.year_slug,
     yearTitle: modifiedProgrammeFields.year_description,
-    lessons: unitLessons,
+    lessons: unitLessons.map((lesson) => ({
+      ...lesson,
+      isUnpublished: true,
+    })),
     pathwaySlug: modifiedProgrammeFields.pathway_slug,
     pathwayTitle: modifiedProgrammeFields.pathway,
     pathwayDisplayOrder: modifiedProgrammeFields.pathway_display_order,
@@ -130,17 +110,18 @@ export const getPackagedUnit = (
   };
 };
 
-const lessonListingQuery =
+const teacherPreviewLessonListingQuery =
   (sdk: Sdk) => async (args: { programmeSlug: string; unitSlug: string }) => {
-    const res = await sdk.lessonListing(args);
+    const res = await sdk.teacherPreviewLessonListing(args);
 
     const modifiedLessons = applyGenericOverridesAndExceptions<
-      LessonListingQuery["lessons"][number]
+      TeacherPreviewLessonListingQuery["lessons"][number]
     >({
       journey: "teacher",
       queryName: "lessonListingQuery",
       browseData: res.lessons,
     });
+
     if (modifiedLessons.length === 0) {
       throw new OakError({ code: "curriculum-api/not-found" });
     }
@@ -154,7 +135,7 @@ const lessonListingQuery =
         ...acc,
         programmeFields: lesson.programme_fields,
         unitSlug: lesson.unit_slug ?? "",
-        unitvariantId: lesson.unitvariant_id ?? 0,
+        unitvariantId: lesson.null_unitvariant_id ?? 0, // it should be unitvariant_id but we don't have it available on this mv
         programmeSlug: lesson.programme_slug ?? "",
         unitTitle:
           lesson.programme_fields.optionality ?? lesson.unit_data.title,
@@ -166,4 +147,4 @@ const lessonListingQuery =
     return lessonListingPageDataSchema.parse(packagedUnit);
   };
 
-export default lessonListingQuery;
+export default teacherPreviewLessonListingQuery;