+++ src/node-lib/curriculum-api-2023/queries/unitListing/unitListing.schema.ts
@@ -50,12 +50,14 @@ const reshapedUnitData = z.object({
   unitStudyOrder: z.number(),
   expired: z.boolean().nullable(),
   expiredLessonCount: z.number().nullable(),
+  unpublishedLessonCount: z.number(),
   yearTitle: yearDescriptions,
   year: yearSlugs,
   yearOrder: z.number(),
   cohort: z.string().nullish(),
   learningThemes: z.array(learningThemesSchema).nullable(),
   subjectCategories: z.array(subjectCategorySchema).nullish(),
+  groupUnitsAs: z.string().nullish(),
 });
 
 export type ReshapedUnitData = z.infer<typeof reshapedUnitData>;
@@ -110,6 +112,7 @@ const unitListingData = z.object({
   subjectCategories: z.array(subjectCategorySchema),
   pathwayTitle: pathways.nullable(),
   relatedSubjects: z.array(subjectSlugs).optional(),
+  groupUnitsAs: z.string().optional(),
 });
 
 export type UnitListingData = z.infer<typeof unitListingData>;