+++ src/utils/curriculum/formatting.test.ts
@@ -24,7 +24,7 @@ describe("getYearGroupTitle", () => {
       expect(
         getYearGroupTitle(
           {
-            ["All years"]: {
+            ["all-years"]: {
               pathways: [],
               units: [],
               childSubjects: [],
@@ -34,7 +34,7 @@ describe("getYearGroupTitle", () => {
               groupAs: "Swimming",
             },
           },
-          "All years",
+          "all-years",
         ),
       ).toEqual("Swimming (all years)");
     });
@@ -64,7 +64,7 @@ describe("getYearGroupTitle", () => {
       expect(
         getYearGroupTitle(
           {
-            ["All years"]: {
+            ["all-years"]: {
               pathways: [],
               units: [],
               childSubjects: [],
@@ -74,7 +74,7 @@ describe("getYearGroupTitle", () => {
               groupAs: "Swimming",
             },
           },
-          "All years",
+          "all-years",
           "units",
         ),
       ).toEqual("Swimming units (all years)");