+++ src/utils/curriculum/sorting.ts
@@ -3,7 +3,7 @@ import { Subject, SubjectCategory, Tier, Unit } from "./types";
 import { Actions } from "@/node-lib/curriculum-api-2023/shared.schema";
 
 export function sortYears(a: string, b: string) {
-  if (a === "All years") {
+  if (a === "all-years") {
     return -1;
   }
   return parseInt(a) - parseInt(b);